Fixed tests not running issue
